+++ src/node/db/PadManager.js
@@ -22,7 +22,6 @@
 const CustomError = require('../utils/customError');
 const Pad = require('../db/Pad');
 const db = require('./DB');
-const hooks = require('../../static/js/pluginfw/hooks');
 
 /**
  * A cache of all loaded Pads.
@@ -126,7 +125,6 @@ exports.getPad = async (id, text, authorId = '') => {
 
   // initialize the pad
   await pad.init(text, authorId);
-  hooks.callAll('padLoad', {pad});
   globalPads.set(id, pad);
   padList.addPad(id);