+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 
 * New `expressPreSession` server-side hook.
 * New `padDefaultContent` server-side hook.
+* The `db` property on Pad objects is now public.
 * New `getAuthorId` server-side hook.
 * New APIs for processing attributes: `ep_etherpad-lite/static/js/attributes`
   (low-level API) and `ep_etherpad-lite/static/js/AttributeMap` (high-level