+++ CHANGELOG.md
@@ -77,6 +77,7 @@
   * `padCreate`: The `author` context property is deprecated; use the new
     `authorId` context property instead.
   * `padLoad`: Now runs when a temporary Pad object is created during import.
+    Also, it now runs asynchronously.
   * `padRemove`: The `padID` context property is deprecated; use `pad.id`
     instead.
   * `padUpdate`: The `author` context property is deprecated; use the new