+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.8.18
+
+Released: 2022-05-05
+
+### Notable enhancements and fixes
+
+  * Upgraded ueberDB to fix a regression with CouchDB.
+
 # 1.8.17
 
 Released: 2022-02-23